+++ packages/terra-list/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+* Changed
+  * Removed heading level `1` for list sections.
+  * Added support for `Home` and `End` key navigation.
+
 ## 4.55.0 - (February 15, 2023)
 
 * Changed